backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0
sqlalchemy==2.0.23
pydantic==2.5.0
python-dotenv==1.0.0
openai==1.12.0
python-multipart==0.0.6
firebase-admin==6.2.0
pytesseract==0.3.13
Pillow==11.3.0
google-cloud-storage==3.4.0
PyMuPDF==1.26.5
dropbox==12.0.2
easyocr==1.7.2
opencv-python==4.12.0.88
requests==2.32.5
google-auth==2.35.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.149.0
python-docx==1.1.2